backend/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.1.0
Authlib==1.3.1
beautifulsoup4==4.13.3
billiard==4.2.1
boto3==1.24.0
botocore==1.27.96
cachetools==5.5.2
celery==5.5.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==44.0.1
dataclasses-json==0.6.7
decorator==4.4.2
Deprecated==1.2.18
dirtyjson==1.0.8
distlib==0.4.0
distro==1.9.0
Django==4.2.23
django-cors-headers==4.7.0
django-redis==6.0.0
djangorestframework==3.16.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
filelock==3.17.0
filetype==1.2.0
Flask==2.2.2
frozenlist==1.5.0
fsspec==2025.2.0
google-ai-generativelanguage==0.4.0
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.43.0
google-generativeai==0.3.2
googleapis-common-protos==1.68.0
greenlet==3.1.1
grpcio==1.67.1
grpcio-health-checking==1.67.1
grpcio-status==1.62.3
grpcio-tools==1.67.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
kombu==5.5.4
llama-cloud==0.1.12
llama-cloud-services==0.6.1
llama-index==0.12.19
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.0
llama-index-core==0.12.19
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.7
llama-index-llms-openai==0.3.20
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.5
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.1
MarkupSafe==3.0.2
marshmallow==3.26.1
milvus-lite==2.4.11
moviepy==1.0.3
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
mysqlclient==2.2.4
nest-asyncio==1.6.0
networkx==3.2.1
newrelic==10.14.0
nltk==3.9.1
numpy==2.0.2
openai==0.27.0
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
Pillow==10.0.0
platformdirs==4.4.0
proglog==0.1.10
prompt_toolkit==3.0.51
propcache==0.2.1
proto-plus==1.26.0
protobuf==4.25.6
psycopg2-binary==2.9.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pymilvus==2.5.4
pyparsing==3.2.1
pypdf==5.3.0
python-dateutil==2.9.0.post0
python-dotenv==0.19.2
pytz==2025.1
PyYAML==6.0.2
redis==6.2.0
regex==2024.11.6
requests==2.31.0
rsa==4.9
s3transfer==0.6.2
safetensors==0.5.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
sqlparse==0.5.3
striprtf==0.0.26
sympy==1.13.1
tenacity==9.0.0
tiktoken==0.9.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typing-inspect==0.9.0
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.10.0
uritemplate==4.1.1
urllib3==1.26.20
validators==0.34.0
vine==5.1.0
virtualenv==20.35.3
wcwidth==0.2.13
weaviate-client==4.11.0
websockets==15.0.1
Werkzeug==2.2.2
whitenoise==6.11.0
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
